Refactors polycommit into algorithms
