+++ Cargo.toml
@@ -28,7 +28,6 @@ members = [
   "gadgets",
   "marlin", # excluded from [features], to be merged in `algorithms` and `gadgets` once stable
   "parameters",
-  "polycommit", # excluded from [features], to be merged in `algorithms` and `gadgets` once stable
   "profiler", # excluded from [features], for internal use primarily
   "r1cs",
   "utilities",
