+++ circuits/Cargo.toml
@@ -87,10 +87,6 @@ version = "1.0"
 path = "../marlin"
 version = "0.7.5"
 
-[dev-dependencies.snarkvm-polycommit]
-path = "../polycommit"
-version = "0.7.5"
-
 [dev-dependencies.criterion]
 version = "0.3"
