+++ circuits/src/models/constraint_converter.rs
@@ -229,13 +229,12 @@ mod tests {
 
         // Marlin setup, prove, and verify.
         {
-            use snarkvm_algorithms::crypto_hash::poseidon::PoseidonSponge;
+            use snarkvm_algorithms::{crypto_hash::poseidon::PoseidonSponge, polycommit::sonic_pc::SonicKZG10};
             use snarkvm_curves::bls12_377::{Bls12_377, Fq};
             use snarkvm_marlin::{
                 fiat_shamir::FiatShamirAlgebraicSpongeRng,
                 marlin::{MarlinHidingMode, MarlinSNARK},
             };
-            use snarkvm_polycommit::sonic_pc::SonicKZG10;
             use snarkvm_utilities::rand::test_rng;
 
             type MultiPC = SonicKZG10<Bls12_377>;
