+++ dpc/Cargo.toml
@@ -72,11 +72,6 @@ path = "../parameters"
 version = "0.7.5"
 default-features = false
 
-[dependencies.snarkvm-polycommit]
-path = "../polycommit"
-version = "0.7.5"
-default-features = false
-
 [dependencies.snarkvm-profiler]
 path = "../profiler"
 version = "0.7.5"
