+++ dpc/src/network/testnet1.rs
@@ -32,6 +32,7 @@ use snarkvm_algorithms::{
     crypto_hash::poseidon::PoseidonSponge,
     encryption::ECIESPoseidonEncryption,
     merkle_tree::{MaskedMerkleTreeParameters, MerklePath, MerkleTreeParameters},
+    polycommit::sonic_pc::SonicKZG10,
     prelude::*,
     prf::PoseidonPRF,
     signature::AleoSignatureScheme,
@@ -58,7 +59,6 @@ use snarkvm_gadgets::{
 };
 use snarkvm_marlin::{marlin::MarlinNonHidingMode, FiatShamirAlgebraicSpongeRng, MarlinSNARK};
 use snarkvm_parameters::{testnet1::*, Genesis};
-use snarkvm_polycommit::sonic_pc::SonicKZG10;
 use snarkvm_utilities::{FromBytes, ToMinimalBits};
 
 use once_cell::sync::OnceCell;
