+++ marlin/src/marlin/universal_srs.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use snarkvm_polycommit::PolynomialCommitment;
+use snarkvm_algorithms::polycommit::PolynomialCommitment;
 
 /// The universal public parameters for the argument system.
 pub type UniversalSRS<F, CF, PC> = <PC as PolynomialCommitment<F, CF>>::UniversalParams;
