--- polycommit/Cargo.toml
@@ -1,96 +1,0 @@
-[package]
-name = "snarkvm-polycommit"
-version = "0.7.5"
-authors = [
-  "Alessandro Chiesa <alexch@berkeley.edu>",
-  "Mary Maller <mary.maller.15@ucl.ac.uk>",
-  "Yuncong Hu <huyuncongh@gmail.com>",
-  "William Lin",
-  "Pratyush Mishra <pratyush@berkeley.edu>",
-  "Noah Vesely <noah.vesely.18@ucl.ac.uk>",
-  "Nicholas Ward <npward@berkeley.edu>",
-  "The Aleo Team <hello@aleo.org>"
-]
-description = "A library for constructing polynomial commitment schemes for use in zkSNARKs"
-repository = "https://github.com/AleoHQ/snarkVM"
-keywords = [
-  "aleo",
-  "cryptography",
-  "blockchain",
-  "decentralized",
-  "zero-knowledge"
-]
-categories = [ "cryptography::cryptocurrencies", "operating-systems" ]
-include = [
-  "Cargo.toml",
-  "src",
-  "README.md",
-  "LICENSE-APACHE",
-  "LICENSE-MIT"
-]
-license = "GPL-3.0"
-edition = "2018"
-
-[dependencies.snarkvm-algorithms]
-path = "../algorithms"
-version = "0.7.5"
-default-features = false
-features = [ "fft", "msm", "crh" ]
-
-[dependencies.snarkvm-curves]
-path = "../curves"
-version = "0.7.5"
-default-features = false
-
-[dependencies.snarkvm-fields]
-path = "../fields"
-version = "0.7.5"
-default-features = false
-
-[dependencies.snarkvm-profiler]
-path = "../profiler"
-version = "0.7.5"
-default-features = false
-
-[dependencies.snarkvm-utilities]
-path = "../utilities"
-version = "0.7.5"
-default-features = false
-
-[dependencies.derivative]
-version = "2"
-features = [ "use_core" ]
-
-[dependencies.digest]
-version = "0.10"
-
-[dependencies.hashbrown]
-version = "0.12.0"
-
-[dependencies.itertools]
-version = "0.10.3"
-
-[dependencies.rand_core]
-version = "0.6"
-default-features = false
-
-[dependencies.rand]
-version = "0.8"
-default-features = false
-features = [ "std_rng" ]
-
-[dependencies.rayon]
-version = "1"
-optional = true
-
-[features]
-default = [
-  "parallel",
-  "snarkvm-algorithms/default",
-  "snarkvm-curves/default",
-  "snarkvm-fields/default",
-  "snarkvm-utilities/default"
-]
-std = [ ]
-parallel = [ "rayon", "std", "snarkvm-algorithms/parallel", "snarkvm-utilities/parallel" ]
-print-trace = [ "snarkvm-profiler/print-trace" ]
