+++ wasm/Cargo.toml
@@ -59,12 +59,6 @@ optional = true
 default-features = false
 features = ["wasm"]
 
-[dependencies.snarkvm-polycommit]
-path = "../polycommit"
-version = "0.7.5"
-optional = true
-default-features = false
-
 [dependencies.snarkvm-r1cs]
 path = "../r1cs"
 version = "0.7.5"
@@ -111,7 +105,6 @@ full = [
   "fields",
   "gadgets",
   "marlin",
-  "polycommit",
   "r1cs"
 ]
 algorithms = [ "snarkvm-algorithms" ]
@@ -120,5 +113,4 @@ dpc = ["snarkvm-dpc"]
 fields = [ "snarkvm-fields" ]
 gadgets = [ "snarkvm-gadgets" ]
 marlin = [ "snarkvm-marlin" ]
-polycommit = [ "snarkvm-polycommit" ]
 r1cs = [ "snarkvm-r1cs" ]
